[PRACTICE-5] 429 에러 발생으로 인한 동시 발송 수 제한 및 외부사 응답 파라미터 바인딩 오류 해소

+++ src/main/kotlin/park/sangeun/aysnctest/domain/exchange/model/CurrencyResultResponse.kt
@@ -1,5 +1,7 @@
 package park.sangeun.aysnctest.domain.exchange.model
 
+import java.math.BigDecimal
+
 data class CurrencyResultResponse(
     val success: Boolean,
     val terms: String,
@@ -16,6 +18,6 @@ data class CurrencyResultResponse(
 
     data class Info (
         val timestamp: Long,
-        val quote: Float
+        val quote: BigDecimal
     )
 }
