+++ src/main/kotlin/park/sangeun/aysnctest/domain/exchange/model/ExchangeRateResponse.kt
@@ -1,9 +1,11 @@
 package park.sangeun.aysnctest.domain.exchange.model
 
+import java.math.BigDecimal
+
 data class ExchangeRateResponse(
     val to: String,
     val from: String,
     val amount: Long,
-    val rate: Float,
+    val rate: BigDecimal,
     val result: Float,
 )
